Renamed implementation_types.h
