+++ libalgebra/implementation_types.h
