+++ libalgebra/libalgebra.h
@@ -55,7 +55,7 @@ typedef unsigned __int64 uint64_t;
 #include <boost/thread/shared_mutex.hpp>
 
 #include "compat.h"
-#include "implimentation_types.h"
+#include "implementation_types.h"
 #include <boost/thread/locks.hpp>
 
 #if __cplusplus >= 201103L
